reimplemented zombie users based on pclient branch changed concept:   recovering a zombie can happen from any server, so we cannot   retain the zombies numnick. Instead I'm using the numnick of   the recovering user for the zombie. The recovering user gets   the numnick of the zombie instead - but the client is killed   in that process anyway.
To make a client a zombie on connection close the user has to
set user mode +Z first. Then the client is no longer removed
from the network (at least for 10 minutes) and can be recovered
using /unzombie <nick>  from another client.
Currently there is no account check or anything to restrict who
is able to recover a zombie.

# Conflicts:
#	include/client.h
#	ircd/s_user.c